requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Caching==2.1.0
peewee==3.17.0
requests==2.31.0
beautifulsoup4==4.12.2
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-api-python-client==2.78.0
python-dotenv==1.0.0
pyttsx3
torch
torchaudio
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
# chatterbox-tts